intermediate/requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
ollama==0.2.1
openpyxl==3.1.2
twilio==9.0.4
python-dotenv==1.0.1
APScheduler==3.10.4
qrcode[pil]==7.4.2
requests==2.32.3
gunicorn==21.2.0
libsql-experimental==0.0.55
